copysec_risk_analyzer/requirements.txt
openai>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
yfinance>=0.2.33 